separate the test for rotation_constraint_internal
